Showing doses by day in main screen from LocalSource

+++ Aurora/app/src/main/java/com/iti/aurora/database/ConcreteLocalSource.java
@@ -70,7 +70,7 @@ public class ConcreteLocalSource implements LocalSource {
     }
 
     @Override
-    public Medicine getSpecificMedicine(int medId) {
+    public Medicine getSpecificMedicine(long medId) {
         return medicineDAO.getSpecificMedicine(medId);
     }
