+++ Aurora/app/src/main/java/com/iti/aurora/model/Repository.java
@@ -47,7 +47,7 @@ public class Repository implements RepositoryInterface{
     }
 
     @Override
-    public Medicine getSpecificMedicine(int medId) {
+    public Medicine getSpecificMedicine(long medId) {
         return localSource.getSpecificMedicine(medId);
     }
