+++ Aurora/app/src/main/java/com/iti/aurora/model/RepositoryInterface.java
@@ -16,7 +16,7 @@ public interface RepositoryInterface {
     @NonNull Single<Long> insertMedicine(Medicine medicine);
     void deleteMedicine(Medicine medicine);
     void updateMedicine(Medicine medicine);
-    Medicine getSpecificMedicine(int medId);
+    Medicine getSpecificMedicine(long medId);
     LiveData<List<Medicine>> getAllStoredMedicines();
 
     @NonNull Single<Long> insetTreatment(Treatment treatment);
